Move BannerModal into its own folder to keep supporting components together.
