+++ assets/js/components/BannerModal/GraphicContainer.tsx
