+++ assets/js/components/BannerModal/index.stories.tsx
@@ -24,7 +24,7 @@ import { ComponentProps } from 'react';
 /**
  * Internal dependencies
  */
-import BannerModal from './BannerModal';
+import BannerModal from './index';
 // @ts-expect-error - We need to add types for imported SVGs.
 import WelcomeModalGraphic from '@/svg/graphics/welcome-modal-graphic.svg';
