+++ assets/js/components/WelcomeModal.tsx
@@ -44,7 +44,7 @@ import {
 import { MODULES_ANALYTICS_4 } from '@/js/modules/analytics-4/datastore/constants';
 import { MODULES_SEARCH_CONSOLE } from '@/js/modules/search-console/datastore/constants';
 import { MODULE_SLUG_ANALYTICS_4 } from '@/js/modules/analytics-4/constants';
-import BannerModal from '@/js/components/BannerModal';
+import BannerModal from '@/js/components/BannerModal/index';
 import { useShowTooltip } from '@/js/components/AdminScreenTooltip';
 // @ts-expect-error - We need to add types for imported SVGs.
 import WelcomeModalGraphic from '@/svg/graphics/welcome-modal-graphic.svg';
