+++ assets/js/modules/analytics-4/components/site-goals/IntroModalBanner/IntroModalEcommerce.tsx
@@ -25,7 +25,7 @@ import { createInterpolateElement } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import BannerModal from '@/js/components/BannerModal';
+import BannerModal from '@/js/components/BannerModal/index';
 // @ts-expect-error - We need to add types for imported SVGs.
 import SiteGoalsIntroModalEcommerceGraphic from '@/svg/graphics/site-goals-intro-modal-ecommerce.svg';
 import Link from '@/js/components/Link';
